# Step 1: Secretly add return_counts to unique, and refactor unique_dim for performance