@VitalyFedyunin 2019-04-02
@pytorch retest this please